# search-api/public/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Arama ve ürün API",
    "version": "1.0.0",
    "description": "PoC: API anahtarı gerekmez. Ürün değişiklikleri app:products:index komutundan sonra aramaya yansır."
  },
  "servers": [
    {
      "url": "/"
    }
  ],
  "paths": {
    "/api/search": {
      "post": {
        "tags": [
          "Arama"
        ],
        "operationId": "search",
        "summary": "Semantik veya hibrit ürün arama",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SearchRequest"
              },
              "example": {
                "query": "tam yatan bebek arabası",
                "mode": "semantic",
                "limit": 10
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Başarılı",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SearchResponse"
                }
              }
            }
          },
          "400": {
            "description": "Geçersiz istek",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "503": {
            "description": "Bağımlı servis kullanılamıyor",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "Product": {
        "type": "object",
        "properties": {
          "id": {
            "oneOf": [
              {
                "type": "string"
              },
              {
                "type": "integer"
              }
            ]
          },
          "variants": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "id": {
                  "oneOf": [
                    {
                      "type": "string"
                    },
                    {
                      "type": "integer"
                    }
                  ]
                },
                "name": {
                  "type": "string"
                },
                "merchants": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "additionalProperties": true
                  }
                }
              },
              "required": [
                "id"
              ]
            }
          }
        },
        "required": [
          "id",
          "variants"
        ],
        "additionalProperties": true
      },
      "Error": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string"
          }
        },
        "required": [
          "error"
        ]
      },
      "SearchRequest": {
        "type": "object",
        "properties": {
          "query": {
            "type": "string",
            "minLength": 1,
            "description": "Boş olamaz; UTF-8 olarak en fazla 2000 bayt."
          },
          "mode": {
            "type": "string",
            "enum": [
              "semantic",
              "hybrid"
            ],
            "default": "semantic"
          },
          "limit": {
            "type": "integer",
            "minimum": 1,
            "maximum": 50,
            "default": 10
          },
          "filters": {
            "type": "object",
            "properties": {
              "brand_id": {
                "oneOf": [
                  {
                    "type": "string"
                  },
                  {
                    "type": "integer"
                  }
                ]
              },
              "category_ids": {
                "type": "array",
                "minItems": 1,
                "maxItems": 50,
                "items": {
                  "oneOf": [
                    {
                      "type": "string"
                    },
                    {
                      "type": "integer"
                    }
                  ]
                }
              }
            },
            "additionalProperties": false
          }
        },
        "required": [
          "query"
        ]
      },
      "SearchResponse": {
        "type": "object",
        "properties": {
          "mode": {
            "type": "string"
          },
          "count": {
            "type": "integer"
          },
          "items": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "id": {
                  "type": "string"
                },
                "score": {
                  "type": "number"
                },
                "document": {
                  "$ref": "#/components/schemas/Product"
                }
              }
            }
          }
        }
      }
    }
  }
}